Move garage.tf.replay_buffer to garage.replay_buffer
Replay_buffer is a fixed size memory to store experience transitions. It
has no tf dependencies. Move it to garage.replay_buffer.

Replace SimpleReplayPool in garage.algos.ddpg with garage.replay_buffer.
