+++ garage/replay_buffer/__init__.py
@@ -3,4 +3,4 @@ This public package contains the replay buffer primitives.
 
 The replay buffer primitives can be used for RL algorithms.
 """
-from garage.tf.replay_buffer.replay_buffer import ReplayBuffer
+from garage.replay_buffer.replay_buffer import ReplayBuffer
